Opt into unstable libcxx ABI and add a custom namespace.
We use a statically linked libcxx and don't expose any of the internal symbols.
This theoretically allows us to namespace everything to avoid accidentally using
the wrong version libcxx. Also, not having to worry about ABI stability allows
libcxx to opt into optimized routines.

But, since libcxx uses CMake and we don't, the mechanism to generate the
__config_site file doesn't exist in GN. Instead, we check in a file that would
have been generated for us by CMake. The file we check in though is a default
with additional configuration. This version of the file makes it so that we opt
into the unstable ABI (and get optimizations) and also namespace everything for
Flutter so collisions are immediately flagged.