[Impeller] Turn off Aiks bounds tracking for filtered SaveLayers. (#49076)
We use this bounds rect for querying the DisplayList rtree when culling.
But when sub-DisplayLists are being dispatched, the DisplayList's rtree
is local and doesn't incorporate parent layer filters. So
sub-DisplayLists are getting culled as if no filters are being applied.

Technically our local bounds tracking in Aiks is actually correct here,
but this is a quick fix to alleviate the problem. I don't know what the
best longterm solution for this is.

(cherry picked from commit ca329dd84207702a95167b8510a55f3e29de8154)