Remove pthread based thread local support.
The thread_local storage class was not available pre-C++11. Even when C++11 was
available, the C++ runtime in versions of iOS <= 9.0 was not capable of
supporting this storage class. So we ended up using pthread directly in that
case. The unique pointer support was added later. Now that the storage class
has been supported on all Flutter platforms for a while, we can remove the
fallback and remove a bunch of code. The ThreadLocalUniquePtr can be removed
too but that can be attempted in a separate migration.