Resize the RenderTarget before creating and passing its Surface to the VirtualDisplay
This change ensures that we first set the default buffer size on the underlying
SurfaceTexture before creating the Surface for it.

This fixes a bug that only occurs on Android devices running 28 or older:
see https://github.com/flutter/flutter/issues/141068.

This also removes some unnecessary init code and cleans up tests.