More correctly name methods regarding event loop that are currently named microtask
Flutter implements the UI isolates message via posting to a UI task
queue. That task queue has a primary and a secondary queue. The

* primary is used for running tasks resulated to framework (e.g. draw frame)
* secondary is used to run Dart event loop messages (e.g. received data on a socket)

The Dart semantics requires running microtasks before processing the
next event loop message.

The way flutter implements by attaching an observer to the secondary
queue. Every time a task from the queue is run, all observers are run
and one of them is going to run pending microtasks.

In some situations the engine pauses the dart event loop. The
terminology used in the code is "microtask" when in reality what it
means is "event loop".

=> This PR changes this terminology to reflect what actually happens.