Sever the //flutter/vulkan dependency in Flutter tester.
The previous code was setting up a proc table, then getting the address of the
proc that was used to the setup that table, then setting up another proc table.
Directly setup the final proc table and don't depend on //impeller/vulkan.

Part of https://github.com/flutter/flutter/issues/143127