[Impeller] Move impeller/image to impeller/playground/image
Deletes usage in Scene of this code, which was the only
non-test location that used it.

This code imports Skia based code to decompress PNG images, and
it's not clear when we'd need it outside of tests at this point.

It is otherwise a pretty thin wrapper around a texture allocation.