[Impeller] Call vkQueuePresentKHR through the ContextVK's synchronized graphics queue wrapper
SwapchainImplVK had been querying the physical device for a queue that
supports presentation on the specified surface.  However, the selected
queue is typically the same VkQueue instance as the graphics queue
provided by ContextVK.

This could cause a thread policy violation where the raster thread is
calling vkQueuePresentKHR while the IO thread is calling vkQueueSubmit
on the same queue during image decoding.

This PR moves the vkQueuePresentKHR call to the ContextVK's graphics
queue wrapper, which serializes access to the VkQueue.