[Impeller] Don't use constants from an extension.
While reading the Vulkan spec for unrelated reasons, I noticed that
VK_REMAINING_ARRAY_LAYERS is part of VK_KHR_maintenance5. Using this
without enabling the feature is invalid usage as documented in
VUID-VkImageSubresourceLayers-layerCount-09243.

It doesn't seem to trip validation (or perhaps our layers are too old).
I had used the constant because it originally seemed convenient. But 1u
is correct as the barrier is inserted on color attachment which can only
have one layer and lever.