Add et run command
The `run` command builds both the host and target engines and then
invokes `flutter run` so that it runs with the custom engine builds.

It is expected that 'et run' be used inside the directory of a flutter
application.

Command line flags passed after `--` will be forwarded to `flutter run`.

Some examples:

`et run` Build the debug variant and runs the app in that mode.
`et run -- --profile` Build the profile variant and runs the app in that mode.
`et run -- --release` Build the release variant and runs the app in that mode.