[Impeller] Make DeviceHolder follow convention for backend specific utilities.
Utilities below the HAL that are backend specific have the VK (in case
of Vulkan) suffix. The only class that doesn't follow this convention
is/was DeviceHolder. This caused some confusion when reading code
leading to wondering what a device holder above the HAL was supposed to
do/mean. Adds the VK suffix to device holder with a docstring. Also adds
the file the source set as it was missing.

No change in functionality. Just a rename of an existing class.