Fix some races in the platform isolate tests
* Do not check IsolateIsShutdown after CreateAndRegisterIsolate in the MultithreadedCreation test.

The test's main thread could run ShutdownPlatformIsolates between a worker thread's calls to CreateAndRegisterIsolate and IsolateIsShutdown.  If that happens, IsolateIsShutdown will return true even though CreateAndRegisterIsolate succeeded.

* Change the is_registered flag to was_registered and do not clear it during isolate shutdown

Tests like MultithreadedCreation need to know whether RegisterPlatformIsolate succeeded so they can determine if the isolate will be shut down by ShutdownPlatformIsolates or if it needs to be shut down explicitly.  If the flag is cleared during shutdown, then the test may see an isolate that was successfully registered and shut down and think that it was never registered.