Disable cleaning up ImageReaders in memory pressure callback
We originally added this cleanup code to work around a Samsung-specific Android
14 bug where after resuming an application any ImageReaders are busted.
According to the Android team what Samsung is doing is a violation of the
"spec".

The fix ended up breaking VirtualDisplay platform views after a suspend/resume
because the surface we pass to the VirtualDisplay is no longer valid after
the resume and we have no way of fixing that.

This PR removes the Samsung-specific hacky fix, restoring the behaviour
of VirtualDisplay backed platform views.

We have an internal bug with Samsung to address the root cause.