A native Android unit-testing harness.
Sets up rules to create an APK that is comprised of solely native code. Existing
executable targets (like GTests) can then use this to run on Android devices
while having access to activities, windows, etc.. This allows for broader test
coverage. Basically, anything that needed an ANativeWindow could only be tested
in an integration test.

Executables that need access to the native activity must provide an
implementation of `NativeActivityMain` that returns a custom subclass of
`flutter::NativeActivity`. The `native_activity_apk` reads like an `executable`
or `shared_library` target. Just one that packages that executable in an APK.

The APK is built using the Android Tools and does not use Gradle. Creating a new
APK after invalidating some code takes ~200ms on my machine. The edit, compile,
run cycle for only a tiny bit worse than testing on the host.

Builds on top of this new infrastructure to create a `GTestActivity` that runs
an existing test suites. This works really well except the GTest suite logs to
`STDOUT` whereas the engine logs to `logcat`. To quickly work around this, a
custom test status listener has been wired up. This only displays the test
results to logcat today but a similar mechanism can be used to talk to the test
runner in the host. I will wire this up in an upcoming patch as there is no
hooks into this from CI right now.

Creates an APK variant of the `impeller_toolkit_android_unittests` harness.