Update fuchsia_libs.gni
### Motivation of the change:

The Fuchsia SDK target //sdk/lib/vfs/cpp is being migrated to a shared library in https://fxrev.dev/981373. The shared library has already been included in the SDK as a dependency, but is missing from the OOT copies of the runners used in tests (shell/platform/fuchsia/flutter/src.gn). This change adds the missing .so to these prebuilts.


### Blocking issue:

As part of an ongoing Fuchsia SDK migration effort, this library requires migration in order to continue ongoing maintenance efforts.  See https://fxbug.dev/293936429 for details.

### Solutions:

This library is already included and licensed under the exsting Fuchsia SDK license.  It can be verified as already existing in the distributed IDKs due to it being added as a dependency early.  The OOT test archives are the last remaining step before switching the library to start using the new .so.

Bug: [b/293936429](https://fxbug.dev/293936429)

FYI: @zijiehe

[C++, Objective-C, Java style guides]: https://github.com/flutter/engine/blob/main/CONTRIBUTING.md#style