[skwasm] Fix `toString` methods on Paint and ImageFilter/ColorFilter
This fixes https://github.com/flutter/flutter/issues/141639

Most of this was previously unimplemented. It turns out the reason for the
hang described in the github issue was that there was a typo in the name of
the `getMiterLimit` C function, so if the client actually called that method
the Wasm module failed to compile, as it couldn't find an import with the
misspelled name.