[et] Fix path canonicalization
Previously we were getting the enclosing directory path of the `et` tool
(via the `dirname` builtin), then canonicalising the path by resolving
symlinks. This doesn't work if the `et` on the path is itself a symlink.
For example, if a user created a symlink `et` in a concrete directory
at `~/.local/bin`:

```
~/.local/bin/et -> ~/src/flutter/engines/src/flutter/bin/et
```
first we'd compute the dirname of the script:
```
~/.local/bin
```
then we'd resolve symlinks:
```
~/.local/bin
```
and incorrectly assume that was the engine directory.

Instead, we now resolve symlinks, then compute the enclosing directory.