[et] Simplify path canonicalisation logic
In order to canonicalise paths, previously we were doing an iterative
computation to resolve symlinks to a canonical path, directory by
directory. This was because on macOS and other BSDs, readlink doesn't
support the `-f` (follow symlinks) option. However, macOS and other
BSD-based systems *do* bundle the `realpath` utility, which resolves
symlinks.