[Impeller] Wait for the previous AHB texture to be fully recyclable.
The Android surface transaction completion handler documentation
states:

```
Buffers which are replaced or removed from the scene in the transaction
invoking this callback may be reused after this point.
```

However, this is NOT sufficient. One also needs to be obtain the
previous release fence and perform a wait for the buffer to be safely
reusable.

This patch adds a GPU side wait for this fence using available
extensions.