[icu] Manual roll of icu to 98f2494518c2dbb9c488e83e507b070ea5910e95
The wording change in the license file in ICU 74.1 required adapting
the license-recognizing regex pattern. This is why the roll is manual
instead of automated.

Similarly, changes in Unicode data files required expanding the
allowlist of patterns which have spurious occurences of the "Copyright"
symbol.

Issue: https://github.com/flutter/flutter/issues/149684