Prepare engine for deprecation of async_minitest.dart
The Dart SDK's `package:async_helper/async_minitest.dart`
was never intended for new tests, it was a drop-in
polyfill for the legacy `package:unit_test` that some
Dart SDK tests were using. The tests should never have used
that package, and migrating them to `package:test` was not
an option, so the `async_minitest.dart` file was the
minimal polyfill that made all the tests run.

Since then, new tests have been written using the file,
and to stop that, the file will now become deprecated,
and (hopefully some day) removed.

If the Flutter engine wants to keep using the file,
they should own their own copy of it, which is what
this change does.

(Also migrate off the to-be-deprecated
`assertStatementsEnabled` property of
`package:expect/expect.dart`, moved to the new
`asserts` from `package:expect/variations.dart`.)