Restore creation of engine before Linux widget is realized.
Due to changes in the renderer in 6b857de the engine was created once
a widget is realized, not when the widget is created. If a Flutter application
changed the default my_application.cc template to show the Flutter widget after
plugins are run then these plugins would not be able to access the engine.

Solved by removing the GdkWindow from the renderer constructor and setting in
later when the widget is realized. This works because the renderer is not
used until the widget is realized.

Fixes https://github.com/flutter/flutter/issues/144873