[icu] Ignores the dir `flutter/third_party/icu/patches`
This directory consistes of patches that Chromium's maintainers
of the Chromium ICU fork have already applied to the ICU
source code.

Usually these are fix-forwards that have been applied to the
ICU library after it has been released.

The patch files are kept to ensure that the patches are applied
if we ever need to create a new branch, or update the major ICU
version.

The patch files specifically do *not* get used during the build
process.

Issues: https://github.com/flutter/flutter/issues/151116