[Impeller] Disable dithering by default.
My expectation was that all toggles were disabled with OpenGL. But the
specification states: "The initial value for each capability with the
exception of GL_DITHER is GL_FALSE. The initial value for GL_DITHER is
GL_TRUE."

The kind of dithering is implementation dependent and our layered OpenGL
implementations choose to do nothing when this is enabled (unless
extensions like VK_EXT_legacy_dithering are supported). Since, again,
per the spec: "As far as OpenGL layering is concerned, no dithering is
technically an acceptable algorithm".

But it does have a cost of older hardware where dithering is supported.

This sets the default to match our other backends and get some
performance on hardware that does do dithering.