Avoid using a private GTEST macro to skip tests.
I admit to using this private macro initially without realizing that the
documented way to do this is to use the stream the other macro returns.
The non-private variant is actually more powerful since it allows for
easier customization of the reason for the skip.

No change in functionality. Just removes the use of the private macro.