feat: log patch load address for later symbolication
This isn't perfect, but at least gives us a way to be able to
symbolicate crashes after the fact on iOS if we have *both* an
ips *and* logs.

Ideally we'd find a way to pass this information through whatever
system logs the ips binary images list, but we probably won't ever
be able to do that (private apis).

Also this should just print one number (the load address) instead
I printed the symbols (since they were convenient).

I was not able to get my local setup working so I don't know
if this successfully prints or not.