Rename FlBackingStoreProvider to FlFramebuffer
The name was confusing, as it only provided one type of backing store. FlBackingStoreProvider
sounds more like an interface that the FlFramebuffer would implement if we had multiple
backing stores.

Use OpenGL types for values.

Fix cases where this was sometimes called a texture.

Improve documentation.

Remove the use of GdkRectangle for dimensions - the framebuffer only has a width and a height.
There was code that was using the x,y values which would always have been zero - this has now
been removed.