Simplify the FlView constructor
Set everything up in the construction function, instead using properties and a
constructed vfunc. While this is the pure GObject way of doing things we
haven't done it in any other classes in Flutter and it is more complex than
doing it all in one function. We no longer need to keep the project ref in the
FlView which is never used beyond construction.