Add SK_USE_LEGACY_BLUR_RASTER staging flag
Skia's CPU-based blur implementation is inaccurate for sigma < 2, but the Dart imagefilter test uses a case with sigma = 1.

In https://skia-review.googlesource.com/c/skia/+/885107, the algorithm is updated to be more accurate when sigma < 2, so the output is more accurate. In order to roll into flutter, the expected color values in the tests will need to be updated with the removal of this staging flag.