[Impeller] Allow jumpstarts to pipeline readiness.
Under normal circumstances, all Impeller pipelines are ready well before
the first frame begins processing. This is true even in cases where
there is no existing pipeline cache. When there is a pipeline cache,
startup is even faster. A device specific pipeline cache is saved after
first-start.

However, low-end devices like the Mokey are... really low-end. Pipeline
setup (via the ContentContext) may be the bottleneck. In such cases, the
observation is that not all pipelines need to be ready for frame
rendering to begin. But the pipelines that are needed for the first
frame must be prioritized before the rest.

This patch adds a job queue where job priorities can be controlled by
augmenting blocking accesses to the pipeline futures by making them bump
up the priority of the pending job in the job queue.

On most devices, this will do nothing as all pipelines will be ready. On
really low-end devices, this will time to first frame on first launch.

Additionally, traces have been augmented to make it more apparent which
pipeline is causing the startup stall. This can aid in reordering
first-frame critical pipelines on low-end devices.