[Impeller] Reorder pipeline construction in content context.
Fixes the regression to the first-frame time on the Mokey.

Also add a trace that should allow for easier identification of problematic pipelines.

A more involved job-queue that allows construction jobs to skip the queue was written in https://github.com/flutter/engine/pull/54355. But the Mokey case is not important enough for the additional complexity. See the context in the linked patch.