[CP-stable]Add a precision to the fragment shader (#54518)
This pull request is created by [automatic cherry pick workflow](https://github.com/flutter/flutter/blob/main/docs/releases/Flutter-Cherrypick-Process.md#automatically-creates-a-cherry-pick-request)
Please fill in the form below, and a flutter domain expert will evaluate this cherry pick request.

### Issue Link:
What is the link to the issue this cherry-pick is addressing?

https://github.com/flutter/flutter/issues/152297

### Changelog Description:
Explain this cherry pick in one line that is accessible to most Flutter developers. See [best practices](https://github.com/flutter/flutter/blob/main/docs/releases/Hotfix-Documentation-Best-Practices.md) for examples

Fix black screen on OpenGL ES, due to lack of precision in fragment shader.

### Impact Description:
What is the impact (ex. visual jank on Samsung phones, app crash, cannot ship an iOS app)? Does it impact development (ex. flutter doctor crashes when Android Studio is installed), or the shipping production app (the app crashes on launch)

On OpenGL ES, a black screen is shown instead of the Flutter output. This is due to the fragment shader failing compilation due to it lacking a precision statement. Note that this is required in the OpenGL ES specification, though some drivers continue to work instead using a default value for precision.

### Workaround:
Is there a workaround for this issue?

There is no workaround.

### Risk:
What is the risk level of this cherry-pick?

### Test Coverage:
Are you confident that your fix is well-tested by automated tests?

### Validation Steps:
What are the steps to validate that this fix works?

1. Create default flutter app (`flutter create hello_world`)
2. Run Flutter app on an OpenGL ES device, e.g. Raspberry Pi (`flutter run`).

Expected result:
Default app is correctly rendered.