iOS: Add dSYM binaries to without_entitlements.txt
In flutter/engine#54414, we added dSYM files for physical and simulator
binaries in both regular and extension-safe framework builds, but did
not add the dSYMs to the without_entitlements.txt list.

This passed all engine pre/post-submit tests, as well as framework
tests, but failed during release codesigning in Cocoon in a test here:
https://github.com/flutter/cocoon/blob/d849b14bab90e0f90e2f7667e37c9f9a5696b918/cipd_packages/codesign/lib/src/file_codesign_visitor.dart#L305-L313

This adds the missing files to without_entitlements.txt.

Issue: https://github.com/flutter/flutter/issues/116493