iOS: Add dSYM binaries to without_entitlements.txt
In flutter/engine#54414, we added dSYM files for physical and simulator
binaries in both regular and extension-safe framework builds, but did
not add the dSYMs to the without_entitlements.txt list.

This passed all engine pre/post-submit tests, as well as framework
tests, but failed during release codesigning in Cocoon in a test here:
https://github.com/flutter/cocoon/blob/d849b14bab90e0f90e2f7667e37c9f9a5696b918/cipd_packages/codesign/lib/src/file_codesign_visitor.dart#L305-L313

This adds the missing files to without_entitlements.txt, which fixes a
codesigning error as seen in this build log:
https://ci.chromium.org/ui/p/dart-internal/builders/flutter/Mac%20Production%20Engine%20Drone/13590/overview

A corresponding change was landed on the flutter-3.24-candidate.1 branch:
https://github.com/flutter/engine/pull/54573

The build associated with that patch correctly completed code signing in
this build:
https://ci.chromium.org/ui/p/dart-internal/builders/flutter/Mac%20engine_release_builder/688/overview

And more specifically, this sub-build:
https://ci.chromium.org/ui/p/dart-internal/builders/flutter/Mac%20Production%20Engine%20Drone/13650/overview

And even more specifically, this build step:
https://logs.chromium.org/logs/dart-internal/buildbucket/cr-buildbucket/8739493904842446705/+/u/Global_generators/Codesign__Volumes_Work_s_w_ir_cache_builder_src_out_release_unsigned_artifacts.zip/codesign_Apple_engine_binaries/stdout

Additionally, this patch adds `sky_utils.assert_valid_codesign_config()`
which fails the script (and thus the build) with an error message if any
file in scope for code signing (i.e. Mach-O binaries) is not listed in
the code-signing config (entitlements.txt, without_entitlements.txt), or
any listed file is not found on disk.

Issue: https://github.com/flutter/flutter/issues/116493